server/requirements.txt
flask
requests
slack_bolt
python-dotenv
pyairtable